Smart_home_System/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
influxdb-client
paho-mqtt
python-dotenv
